Moves all assertions tests to the assertions module (#500)
Moves all assertions tests to the assertions module

As part of https://github.com/kotlintest/kotlintest/issues/498, all tests that test the `kotlintest-assertions` module were moved to it, bringing the tests closer to their targets, and decreasing confusion to future contributors.

An important point to be noted: the root package chosen is `assertions.io.kotlintest`. This was done because putting files in `io.kotlintest` package make them subject to stacktrace cleaning, by the Failures class. The Failures class removes all Kotlintest stacktraces from throwables, so that the user only sees the real problem, and nothing internal. However. as we're using `io.kotlintest` for all the tests, the user stacktraces would be removed (we are the users), and that's not good when a test fails, as no reason would be given.

Fixes a part of #498 - Moves tests closer to their targets